Pckgs update, changed psam with phpstan (#124)
### Changes:
- ImageFile: now have new filed `manipulatedAt`, which is updated after rotate image, crud on image rois or reprocess optimal crop.
- Elasticsearch reindex: add option `--since`  which is supported by entity `Asset`
- Logging imporvements: introduced journalLogger, improved exception logging by appLogger
- Added new entity `PodcastExportData` which holds data for combination of DeviceType and ExportType. Entity also works with Anzutap body format

### Dependency / Version Changes
Runtime
anzusystems/common-bundle: ^9.2 → ^10.3
anzusystems/anzutap-bundle: added ^0.10
Symfony core components: ^7.2 → ^7.3
(framework-bundle, dependency-injection, doctrine-bridge, http-kernel, http-client, messenger, validator, security-bundle, etc.)
symfony/monolog-bundle: added & pinned to ^3.11 (compatibility with mongo handler)
symfony/var-exporter: added ^7.3
### Dev / Tooling
Removed: vimeo/psalm, psalm/plugin-symfony, weirdan/doctrine-psalm-plugin
Added:
phpstan/phpstan: ^2.0
phpstan/phpstan-doctrine: ^2.0
phpstan/phpstan-symfony: ^2.0
Upgraded:
phpunit/phpunit: ^9.5 → ^12.0
rector/rector: ^1.0 → ^2.0
Kept: symplify/easy-coding-standard: ^12.1, slevomat/coding-standard: ^7.0

+++ src/DependencyInjection/ByteSizeEnvVarProcessor.php
@@ -44,8 +44,7 @@ final class ByteSizeEnvVarProcessor implements EnvVarProcessorInterface
             return (int) $byteValue;
         }
 
-        /** @var int $value */
-        $value = substr($byteValue, 0, -2);
+        $value = (int) substr($byteValue, 0, -2);
         $unit = strtoupper(substr($byteValue, -2));
 
         if (isset(self::MULTIPLE_MAP[$unit])) {
